pkp/pkp-lib#4787 upgrade migrations moved to v3_6_0 folder
